requirements.txt
lark~=1.1.8
docutils~=0.20.1
pip~=23.2.1
wheel~=0.41.2
Sphinx~=7.2.6
Jinja2~=3.1.2
requests~=2.31.0
Babel~=2.14.0
setuptools~=68.2.0
yarg~=0.1.9
MarkupSafe~=2.1.3
packaging~=23.2
imagesize~=1.4.1
urllib3~=2.1.0
snowballstemmer~=2.2.0
Pygments~=2.17.2
certifi~=2023.11.17
docopt~=0.6.2
pipreqs~=0.4.13
recommonmark~=0.7.1
idna~=3.6
alabaster~=0.7.13
commonmark~=0.9.1
sphinx_rtd_theme
pyyaml